requirements.txt
Flask
torch
Pillow
open_clip_torch
scikit-learn
pandas
numpy
